chore: move component files
