+++ src/modules/feature-panel/components/FeatureDescription.tsx
