+++ src/modules/feature-panel/components/HeaderImageSection.tsx
