+++ src/modules/feature-panel/components/OSMInfoSection.tsx
@@ -1,17 +1,23 @@
 import { Box, Grid, Text } from "@radix-ui/themes";
 import React from "react";
-import TagGroupRenderer from "~/modules/edit/components/TagGroupRenderer";
 import type { TagOrTagGroup } from "~/modules/edit/hooks/useOsmTags";
+import TagGroupRenderer from "~/modules/feature-panel/components/TagGroupRenderer";
+import { TagList } from "~/modules/feature-panel/components/TagList";
+import { TagListItem } from "~/modules/feature-panel/components/TagListItem";
 
 interface Props {
   tags: TagOrTagGroup[];
 }
 
 const OsmInfoSection = ({ tags }: Props) => {
+  console.log("OsmInfoSection tags", tags);
   return (
-    <ul style={{ listStyle: "none", padding: 0, margin: 0 }}>
+    <TagList style={{ listStyle: "none", padding: 0, margin: 0 }}>
       {tags?.map((tagOrTagGroup) => (
-        <li key={tagOrTagGroup.key} style={{ marginBottom: "var(--space-3)" }}>
+        <TagListItem
+          key={tagOrTagGroup.key}
+          style={{ marginBottom: "var(--space-3)" }}
+        >
           <Grid columns="6rem auto" gap="3">
             <Box>
               <Text size="3" color="gray">
@@ -25,9 +31,9 @@ const OsmInfoSection = ({ tags }: Props) => {
               />
             </Box>
           </Grid>
-        </li>
+        </TagListItem>
       ))}
-    </ul>
+    </TagList>
   );
 };
