+++ src/modules/feature-panel/components/PartOf.tsx
